Create a noop OSS logger and a facebook-only logger for external carbon connections
Summary:
This diff adds a logger for external carbon connection.

It is currently turned off, and will have a followup diff with settings to adjust the rate at which we log, as well as an overall on/off switch.

Due to the logger containing facebook only code which is included at compile time, there are two versions of the logger.

The first logger does nothing, and merely contains the required code to allow the connection implementation to call a log function.

The second logger uses the facebook only logger to log various info about each connection. When enabled, it will log a limited set of samples to scuba.

Reviewed By: lenar-f

Differential Revision: D38225565

fbshipit-source-id: 1965b70dbb7cf00ab545ac53b28d95aca358f5ab