add TypeTagToTType to Encode.h
Summary:
Added TypeTagToTType to detail/Encode.h, which converts TypeTag to TType.
This is useful for op::encode and op::decode.
We cannot use type.Type as we cannot include `"thrift/lib/cpp2/type/Type.h"` to this file as it will be used by `module_types.tcc` and creates circular dependency.

Reviewed By: Mizuchi

Differential Revision: D39005964

fbshipit-source-id: b1215a3dc728296f554ddf70a42f687f37e16900