Put most DirectDeclSmartConstructors fields behind an Rc
Summary:
The FFP frequently clones its SmartConstructors state in order to create checkpoints for backtracking. This means that expensive clone operations can really slow down FFP perf.

This state struct has grown to be pretty large. On some degenerate cases, we spend over 90% of parse time cloning! Put most of its fields (all except the token_factory, which is zero-sized, and the previous_token_kind, which is small and changes very frequently) behind an Rc so that cloning is cheap again.

Reviewed By: edwinsmith

Differential Revision: D39038277

fbshipit-source-id: c4cc49e9bd1f9d69020cf148f4f88d30b86141b5