Make class pointer module boundary violations consistent between interp and jit
Summary: Class pointer and class name related module boundary violations were coming from different places in the interp and jit, namely in the interp `ResolveClass` and `ClassGetC` were responsible; whereas, in the jit, `NewObj` was responsible. This is not good as `ResolveClass` and `ClassGetC` can be used on their own without `NewObj`. Lets make the behavior consistent.

Reviewed By: mdko

Differential Revision: D39062685

fbshipit-source-id: f11c1e7e55e388f9e1761ca7491b62c369d9e622