no longer use actions in showStatusRequest
Summary:
The original design of the window/showStatus request in Nuclide was that HackLSP sends status, and it might include actions. For instance if hh_server were stopped then it might include the action "restart hh_server".

I made window/showStatus be a request just so that, if the user clicks on this action, then a response could come back. It was a janky design and should have been done differently (namely, by having window/showStatus be a notification, and any actions in it give rise to VSCode issuing a workspace/executeCommand request). But Nuclide didn't have workspace/executeCommand at the time...

But when should Nuclude even respond to the showStatus request? This was the design:
1. When Nuclide receives a showStatus request with actions, it displays status and action buttons, but doesn't send a response.
2. If the user clicks an action, then it removes the status display, and sends a response.
3. But if the LSP server sends a subsequent showStatus request, then Nuclide should respond to the *prior* showStatus request, basically just closing it out.

The status feature got ported from Nuclide to VSCode. But the port changed behaviors 1+3 in the case where there were no actions. It changed it to respond immediately in these cases. This caused hack to get into an infinite showStatus cycle T54104285. The fix we decided at that time was for VSCode to go back to match Nuclide's behavior D17873178.

However the fix wasn't quite right. It changed VSCode to *never* send responses to actionless status. More recently, VSCode was changed again to send responses per the original Nuclide behavior D37948882. This most recent change surfaced a concurrency bug in hack...

## The bug in hack

Hack uses a global variable `showStatus_outstanding`, intended to show what message it thinks VSCode is currently displaying. The idea is that if hack tries to send the same status twice, it sees that the second status is the same as the first and so just doesn't bother showing it.

To match Nuclide behavior (which is that status gets dismissed when a response arrives), we obviously have to reset `showStatus_outstanding` when that response arrives.

But this suffers an interleaving bug:

1. Hack sends s#6 "ready" and sets `outstanding := "ready"`
2. Hack sends s#7 "working" and sets `outstanding := "working"`
3. VSCode sends a response to s#6, but hack is working and so can't handle it
4. Hack finishes its work and sends s#8 "ready" and sets `outstanding := "ready"`
5. Hack handles response to s#6. At this point, the response to s#6 overwrites the outstanding associated with s#8.

Here's a full transcript: P524661318 (LSP logs), P524661300 (clientLsp logs)
```
10:35:42.884 LSP sends s#2 "Hack IDE: initializing.". outstanding:="initalizing"
10:35:42     VSCode receives s#2
10:35:46.158 LSP sends s#3 "Hack IDE: initializing. hh_server: ready". outstanding:="initializing/ready"
10:35:46     VSCode receives s#3
10:35:46     VSCode responds to s#2
10:35:46.358 LSP handles s#2 response. No change to outstanding
10:35:47.274 LSP sends s#4 "Hack IDE: ready. hh_server: ready". Outstanding="ready/ready"
10:35:47     VSCode receives s#4
10:35:47     VSCode sends response to s#3
10:35:47.282 LSP sends s#5 "Hack IDE: processing. hh_server: ready". outstanding:="processing/ready"
10:35:47     VSCode receives s#5
10:35:47     VSCode sends response to s#4
10:35:47.398 LSP handles s#3 response. No change to outstanding.
10:35:47.681 LSP handles s#4 response. No change to outstanding.
10:35:50.034 LSP sends s#6 "Hack IDE: ready. hh_server: ready". outstanding:="ready/ready"
10:35:50     VSCode receives s#6
10:35:50     VSCode sends response to s#5
10:35:50.438 LSP handles s#5 response. No change to outstanding.
10:36:02.065 LSP sends s#6 "Hack IDE: working. hh_server: ready". outstanding:="working/ready"
10:36:02     VSCode receives s#7
10:36:02     VSCode sends response to s#6
10:36:08.533 LSP sends s#8 "Hack IDE: ready. hh_server: ready." outstanding:="ready/ready"
10:36:08     VSCode receives s#8
10:38:08     VSCode sends response to s#7
10:36:08.543 LSP handles s#6 response "ready/ready".

Even though VSCode sent response to s#6 at 10:36:02, hack was busy at that time "IDE working" and only became ready at 10:36:08. At that time it sent a new status s#8, and then proceeded to handle the response to s#6. But the response to s#6 trampled over the "outstanding" state.
```

## What this diff does, and why

The whole messy reason for request/response was as follows:
1. The status+buttons in Nuclide were very prominent, and visible, and were a natural and nice UX by which users could restart hack without having to go to the command-line
2. Nuclide didn't support workspace/executeCommand, so my yucky workaround was to treat those buttons as responses to the status request. I regretted that choice at the time, and still do! It comes with too much baggage -- namely, (1) having to respond to a previous request only when the subsequent request arrives, (2) dismissing status when the user clicks a button, (3) two separate occasions, T54104285 and S290873, where we've had showStatus-related spew.
3. VSCode doesn't have the nice UX for status. It instead moves status into the status-bar. The actions are almost invisible. They only show up when the user clicks the status indicator. I personally don't believe that anyone does.

I decided to just rip out the entire "status actions" from Hack. We don't record telemetry on whether anyone uses them, but my hunch is (1) no one does, (2) even if they did, it would serve users well to learn the command-line instead. Also this is a buggy area, prone to complexity and mis-interpretations, and it's not worth it.

I could have written this differently, to just no longer clear out `outstanding` upon receipt of a response. This would also solve the bug in hand, and be a slightly more accurate model of what VSCode does. However it wouldn't be a completely correct model of what VSCode does -- since VSCode retains the status even after you click on an action, and indeed lets you click on an action twice if you're so inclined. I think the request/response paradigm is incapable of properly modeling VSCode's behavior. I worry that leaving it in place will be a bug farm. That's why I'm ripping out actions entirely, i.e. simply remove hack from exposure to this bad area.

What this diff does:
1. It no longer uses actions in showStatus.
2. It correspondingly has a much simpler model of VSCode. The model is now simply *"The latest showStatus that we sent is sticky, and will remain visible until we send a new showStatus"*.

Reviewed By: mroch

Differential Revision: D39061012

fbshipit-source-id: c3cb76f6f38c25a011b8aaa21ab4e04ec6cc6532