Verify output over test/decl test cases
Summary: Add a mode to hh_single_decl to compare the output of ocamlrep_marshal to OCaml's Marshal. Add a test running it over the test cases in hphp/hack/test/decl. Add a commented-out test running it over the test cases in hphp/hack/test/typecheck (we can uncomment this once it's better-behaved).

Reviewed By: shayne-fletcher

Differential Revision: D39011810

fbshipit-source-id: bc7e5f78bb37ad5e12beb940064b425a0bb9ec3c