Print unresolved types as _
Summary:
If we can't resolve a type variable, display it as "_" rather than "[unresolved]". This makes type errors easier to read.

For example, if a user writes `await 1` we used to show:

```
-> Expected Awaitable<[unresolved]>
-> But got int
```

The problem isn't that we couldn't find an inner type for `Awaitable`. The problem is that `1` isn't an awaitable value at all! We now show this:

```
-> Expected Awaitable<_>
-> But got int
```

This is much easier to read, especially in larger types. It's also consistent with `Terr`.

Reviewed By: hgoldstein

Differential Revision: D39067882

fbshipit-source-id: 0e758cf36fe9c73f3b6347d7f53572fabeadeca3