Set clientCAFiles instead of clientCAFile
Summary:
Splitting up D37429044

We are deprecating the clientCAFile field in SSLContextConfig in favor of the clientCAFiles field. Both are used for the same purpose, but clientCAFiles was added to support multiple ca files in a vector. clientCAFiles is already being used for the majority of use cases after D38295264 updated ThriftStructProcessor.cpp, but there is a long tail of custom configured SSLContextConfig's that need to be updated.

Reviewed By: xybu

Differential Revision: D39045696

fbshipit-source-id: 44975bc6070d93f245dd3048bb47766b3cf2d182