print chain of HhError
Summary:
I was struggling with a test failure which merely printed "Unexpected: change".

The reason was that the printer for HhError was hiding all the details of the anyhow it contains. It only printed the outermost cause on the chain, in this case "change".

With this diff, the printer for HhError now prints everything, e.g. "Unexpected: change: Naming table lookup failed: UNIQUE constraint failed: NAMING_SYMBOLS.HASH: disabled backtrace: StringifiedError"

I also noticed that the anyhow context string "change" was used in two places, so I renamed one of them to "change-request".

Reviewed By: edwinsmith

Differential Revision: D38889444

fbshipit-source-id: 804d0c20f716cfbc2e8bb2f0d007c595c0a70cd0