redesign PYI generator to enable cache support for a faster compilation
Summary: Issue this diff is solving is to improve thirft files compilation times for PYI code generator. To improve compilation times we can employ caching approach which python and py3 code generators already utilize. This diff is a refactoring of a PYI generator to adopt python and py3 generators approaches and get caching for free.

Reviewed By: junzh0u

Differential Revision: D39027194

fbshipit-source-id: 8ed0804b72f5f038f5079fdb9f1672d2ccca44ad