Disambiguate code gcc claims is ambiguous.
Summary: gcc claims that this code is ambiguous: we could be constructing an AliasClass and invoking the HPHP::jit::canonicalize that takes an HPHP::jit::AliasClass, or we could be constructing a MemEffects (from a PureLoad or a ReturnEffects, which are implicitly constructable from an AliasClass) and calling the HPHP::jit::canonicalize that takes HPHP::jit::MemEffects. Explicitly constructing an AliasClass in these places makes gcc happy, so do that.

Reviewed By: ricklavoie

Differential Revision: D39028680

fbshipit-source-id: cc8d76b4138810672d2a3a345dc97b1a58feae91