systematically remove the `--extra-builtins` pessimised hhis when running validate_pessimisation
Summary:
The `validate_pessimisation` test suite was not deleting the pessimised hhi files passed via the `--extra-builtins`.  In some cases (eg. on devserver) this could result in the `validate_pessimisation` test suite picking up stale files.

This diff ensures that the `--extra-builtins` pessimised hhis are generated in a unique directory and deleted once the test suite completes.  Directory must be unique because the `pessimise_and_single_type_check.py` script might be invoked concurrently by buck: the <PID> is added to the pessimised extra-builtins hhis directory to avoid race conditions.

The drawback is that manually reproducing a test that relies on extra builtins  requires a manual pessimisation of the extra builtins.

Reviewed By: andrewjkennedy

Differential Revision: D39131926

fbshipit-source-id: 94e8c08e454b635f0815e9a3f1202e41ace4dce2