Handle type accesses on Thas_type_member
Summary:
The invariant that I believed held does not hold. The
test case I introduce here triggers the supposedly
unreachable code. Explanations of how we get there
are in the example but the gist of it is that subtyping
decomposes to the *right* before looking at the *left*.
Because of this, it is possible that Thas_type_member
appears in a subtyping query against a type variable.

We fix the bug by simply implementing the lookup of
type constants in Thas_type_member.

Reviewed By: losvald

Differential Revision: D39097857

fbshipit-source-id: 3f91653031f9be288dc71de727d3f6b4062c1d54