Fix generic constraints involving refinements
Summary:
Having a refinement as the rhs of an `as` constraint on
a generic yielded an invalid tpenv and ignored the
constraint. That's surprising, and unsound. The fix is
simple: when we're adding constraints (require_soundness
boolean set to false), we do not decompose refinements
on the rhs of subtype queries.

This fixes the broken tests introduced by Leo in D39092141 (https://github.com/facebook/hhvm/commit/833d59bbdafb161589ac80e5184cade7be1b1d0d).

Reviewed By: losvald

Differential Revision: D39097858

fbshipit-source-id: 93c689ea8d8389e6bb2cb6e1001ddfe53077b4d3