Avoid dropping uninitialized Vec
Summary:
This assignment can fail nondeterministically depending on what the uninitialized memory happens to be. If it looks like an empty Vec (with no heap-allocated storage), then everything is fine. If it happens to look like a Vec with heap-allocated storage, then Rust attempts to free that storage. This can result in errors like `munmap_chunk(): invalid pointer` or a `SIGBUS`.

Instead, write to the uninitialized memory without attempting to drop its previous contents.

Reviewed By: shayne-fletcher

Differential Revision: D39140665

fbshipit-source-id: 16f301e06f2e0ffff6517c4f4b280e43ad4092f1