Treat internal classes in other modules as opaque
Summary:
This diff makes the typechecker treat any type whose class is internal as opaque outside of the module. We ban most instances of this appearing randomly, but there are ways in the type system to create such cases where the type appears. In those cases, we should treat it like an opaque type for which we know nothing about the value.

I made sure that for most error messages we give a good reason as to why this type looks opaque from outside the module, e.g:
```
  But got `mixed` because this is an internal symbol from module foo, which is opaque outside of the module.
  File "unknown_class.php--file1.php", line 13, characters 12-15:
  The type originated from here
  File "unknown_class.php--file1.php", line 10, characters 11-11:
    via this generic `T````
```

Reviewed By: oulgen

Differential Revision: D38931896

fbshipit-source-id: 4a004f91e9018823dd8c92558f31f42167b5b1c8