Cyclic base types for enums are not checked
Summary:
We currently only perform cycle detection on enum constants, but the
following code would make the runtime fatal

```
enum E : E {}
```

See T128241203 for the full report.
This diff implements this detection for enum (and enum class while we're
at it, even if the runtime would not fatal in that case).

Reviewed By: chenmela

Differential Revision: D39129381

fbshipit-source-id: 9b586e1f5b942d514ddd3a3bd1f07428616413b7