Improve memory effects for ReleaseShallow
Summary:
ReleaseShallow frees a heap object and logically kills its memory locations.
We can't mark it as having kill effects as we have no way to be precise about
them, and we still need to prevent rematerializations based on the memory from
occuring.  The common solution here is to mark it as may store to a superset
of its kill set.  This is then ignored in load elim by an explicit check.

Additionally so that ReleaseShallow can be marked as not reading from the heap,
lets not load the context inside Closure's releaseShallow helper (that isn't a
very shallow thing to do).

Reviewed By: ricklavoie

Differential Revision: D36981040

fbshipit-source-id: b9dc1bbc759191fbc2c74d008cdf651a4a2af9a3