Handle stmem in heap object dce
Summary:
Today we cannot DCE ObjectData's that aren't closures (unless they have no
properties).  This is because we must initialize the properties for a DecRef
(as the DecRef will read them), and the current DCE handling does not handle
LdPropAddr, StMem pairs which is how we will be initializing the dead props.

Reviewed By: ricklavoie

Differential Revision: D39133831

fbshipit-source-id: d3a0eefcc133441e67b0d1f39d7a801992df2d2d