Refactor union find to have a sequential variant
Summary:
The existing union find datastructure assigns sequential ids to the universe.
This is helpful if your data is *not* already sequentially indexed.  If your
data already has indexes it is a waste to recompute such a mapping.  By pulling
the sequential component out, we can make use of that for such cases and avoid
the hash table lookup of the assigned index.

Reviewed By: ricklavoie

Differential Revision: D39133473

fbshipit-source-id: 5c61b1b8799f937563d57333f7585919810a13e0