naming_sqlite::SymbolItem -> summary::SymbolDecl
Summary:
I see that naming_sqlite.rs defines type SymbolItem. But it fills the exact same role+purpose as summary::SymbolDecl: a store of
* string name
* name_type
* decl_hash

I've switched over to SymbolDecl.

The reason I'm doing this is because there's a different data-structure that naming_sqlite needs to have which I'd really like to call naming_sqlite::SymbolItem. It would be a straight-up representation of a row in the main table or the overflow table.
* name_hash
* canon_name_hash
* decl_hash
* file_info_id

Reviewed By: bobrenjc93

Differential Revision: D39122243

fbshipit-source-id: 4bec5cb0802760741b2f26f8e8d1446e433595fe