naming_sqlite remove import crate::datatypes::*
Summary:
I don't like wildcards `use crate::datatypes::*` because whenever the code refers to a type, I don't know where it's defined. So this diff removes it.

I could have left it as `use crate::datatypes` so that the code would look like this:
```
(
        ToplevelSymbolHash,
        datatypes::ToplevelCanonSymbolHash,
        ...
)
```
But really it scared me to see a primitive universal type like "ToplevelSymbolHash" alongside a very crate-internal type "ToplevelCanonSymbolHash", and I thought in this case it was sufficiently worth distinguishing the two with a crate:: prefix

Reviewed By: bobrenjc93

Differential Revision: D39122377

fbshipit-source-id: b2f08b130291f15fe2776449824714a89ac065dc