names/datatypes.rs shouldn't re-export ToplevelCanonSymbolHash
Summary:
I just don't like it when a crate re-exports a type from a different crate. It introduces confusion -- where does the type properly belong? In this case, is ToplevelCanonSymbolHash a naming_sqlite.rs wrapper type for sql, or is it the same as hh24_types::ToplevelCanonSymbolHash?

So I'm removing the "pub use"

Reviewed By: bobrenjc93

Differential Revision: D39137831

fbshipit-source-id: 944b8ac09a4e1ef5d18b150309a182e767772989