Simplify how a user specifies arbitrary quantiles
Summary: This simplifies the process of specifying a specific set of quantiles. A user will still need to wrap an `std::array` inside of a struct, but the requirement to individually label the size is no longer there. Also, this adds a large selection of predefined quantiles that will cover many use cases. A predefined set of quantiles for percentiles is deliberately left off because the cost of inserting a value into a QuantileHistogram is proportional to the number of quantiles.

Reviewed By: guangli-dai

Differential Revision: D38876254

fbshipit-source-id: 774beb5aafac98ba20d8bcc2c746d118b10a0857