Fix fanout bug with required XHP attributes and more complex hierarchies
Summary:
Previously, adding a required XHP attribute to an XHP class would mark the constructor of that class as modified in order to recheck all XHP instances of that class.
Now, we mark the constructor as "Changed_inheritance" instead.
Marking a member m as "Changed_inheritance" is equivalent to marking m as "Modified" in both the class and all its descendants.
That way, we force rechecking all the XHP instances of all the descendants of the modified class.

Reviewed By: bobrenjc93

Differential Revision: D39205926

fbshipit-source-id: 6eda66033f380527a94af283c53225c74471ea1e