use naming_special_names to hold the special name definitions
Summary:
We have a bunch of similarly defined strings in the IR and in naming_special_names - but they're slightly different.

For the ones missing a leading `\HH` I added constants to naming_special_names.
I chose to match the style already in naming_special_names of having the leading backslash so it would be at least somewhat consistent.

Reviewed By: edwinsmith

Differential Revision: D39184055

fbshipit-source-id: 8593ffc86f68d115e5ed3c60df1ecb2ecbf8323f