Fix SP handling with CheckSurprise* and functionEnterHelper
Summary:
CheckSurprise* opcodes call functionEnterHelper, which invokes the intercept
handler. If the function was intercepted, it unwinds the frame and jumps
directly to the saved rip, bypassing the rest of the CheckSurprise* logic.

This is a nice hack that saves code, but it doesn't work if vasm spills into
the stack across the CheckSurprise* logic, as the functionEnterHelper is not
aware of the stack usage. It is not currently possible to have multiple
recordbasenativesp{} regions prevent spilling across CheckSurprise* logic,
while allowing it both before and after.

The next diff will be invoking intercept handler between DV initializers and
jump to the function body, potentially requiring spilling both before and
after.

Fix this by returning back to the CheckSurprise* logic, allowing vasm to clean
up stack before jumping to the rip.

This unfortunately adds some cold code per every function, but should be
negligible compared to the amount of code produced due to inlining.

Once the DV initialization logic is cleaned up, the interception will happen
only at the beginning of the main translation and this hack can be restored.

Reviewed By: ricklavoie

Differential Revision: D39073597

fbshipit-source-id: 31aabc28c29a3c2da8b268372b6ed58c6f4b7eb5