serialize arena_collections::List as a sequence
Summary:
I'd like to be able to compare oxidized_by_ref and oxidized versions of
the same data structure (direct_decl_parser::ParsedFile), by dumping
them both to JSON and running diff.

For that to work, arena_collections::List need to serialize as a sequence
in the same order as Vec.

Reviewed By: shayne-fletcher

Differential Revision: D38912184

fbshipit-source-id: 7bf434bbd4dc582f8e56bbd16bf3944c3a9b9a8f