Delete no longer necessary jit pessimization.
Summary: When const-folding in the jit, we don't fold Eq and Neq comparisons on strings when they're both 'num'-ish but of different types (i.e. one is an int and one is a float). This is no longer needed, since StringData::equal never does a cast to num as part of the comparison any more - its behavior is now identical to StringData::same. Always allow folding Eq and Neq operations on 2 strings.

Reviewed By: paulbiss

Differential Revision: D39269626

fbshipit-source-id: 421032e765d1ddc2ddde7cac23fbbc85dab718c4