Always return the same AAST on parsing errors
Summary:
Previously, the AAST returned on parsing errors would depend on the `fail_open` flag. We also weren't consistent about where we set this flag.

This meant that we discarded large parts of the AAST even in the IDE, resulting in fewer errors reported. Specifically anywhere we called `missing_syntax_` with a replaced node was discarded, and anywhere we saw a missing syntax otherwise would discard that definition or class element entirely.

Instead, always return the same AAST, and remove this flag.

Reviewed By: aorenste

Differential Revision: D39107688

fbshipit-source-id: 3477017b94c8c43ff4b28b124e48cb31abfe52e9