.gitignore for .vscode/tasks.json
Summary:
Previously vassilmladenov landed D18979905 for .vscode/launch.json. This provided a common VSCode>Debug>Hack(dune) for everyone on the team. When you click on the debug button, it pops up a picker inviting you to chose which of two binaries "hh_single_type_check" or "hh_parse" you wish to debug.

Previously vassilmladenov landed D16905948 (https://github.com/facebook/hhvm/commit/c9f841a4774e50885b2fdb5d9735dd22bc4557e5) for .vscode/settings.json for everyone on the team. This way, whenever anyone on the team writes code in VSCode, it standardizes to 2-space indent, and insert-trailing-newline-on-save.

In this diff I am instead adding .vscode/tasks.json to .gitignore. That's because I don't believe there exists a good universal tasks.json that's suitable for everyone on the team, and because I want to allow people on the team to use their own.

tasks.json is used to populate the VSCode Terminal > RunTask/RunBuildTask menu (Cmd+Shift+B). The idea among most OSS projects is developers habitually use Cmd+Shift+B to build within VSCode, they pick which of the available targets they want to be the default for their build, it invokes the build command in the terminal window, and it populates the "Problems" pane with what it gets by scraping the output of the terminal window.

I really want to use this for Rust. Sometimes also for ocaml.

The reason I don't think there exists a good default for everyone on the team is because we have a HUGE number of different possible targets. I imagine that each developer will want to have a fairly simple build menu which only offers to build the few targets they're interested in at the time.

My own tasks.json is here for instance - P527026249

Reviewed By: vassilmladenov

Differential Revision: D39239342

fbshipit-source-id: 0c7ba9bc2a17dab33d3cc8720cd0a49440c32216