Add support for ocamlrep(skip)
Summary:
A future diff needs to be able to skip fields when passing data to/from ocaml.  This adds support to the ocamlrep_derive macro to allow this.

For ToOcamlRep we simply skip the fields.  For FromOcamlRep we initialize skipped fields with Default::default().

Added some tests which show examples.

Differential Revision: D39114410

fbshipit-source-id: 92fdbb4f10775d31c995483121fe9b9af8e5986f