Error on async lambdas
Summary: Discovered that we don't error on async lambdas within expression trees. This was a bit surprising to me as we do not accept `await` statements so I suspect our intention was not to accept async lambdas. There's nothing inherently wrong here about allowing async lambdas - they only affect the return type of the function in question, but we don't actually desugar this in a way that informs the framework of this fact, so it still feels broken.

Reviewed By: benski, Wilfred

Differential Revision: D38060902

fbshipit-source-id: 265ae5999d012eea625882c4ee56762edf8aba39