Fix problem with pool dying while cleaning up
Summary:
While working on the sync connection pool I found another problem with the code that cleans up the pool.  We already had code in to try to handle this correctly.  When the first connection is being cleand up we got its weak pointer to the pool and converted it to a shared pointer in order to guarantee that the pool didn't destruct while we went through the rest of the loop.  However, this doesn't work perfectly.  There is a tiny window of time where if the last owner of the pool releases it right as the cleanup timer fires.  In this scenario, the pool hasn't yet cleaned up but there is no valid shared pointer to it and when the first connection attempts to convert its weak pointer to a shared pointer it gets nullptr which causes a fatal error.

After thinking about it a bit I realized that a better solution is to not cleanup the connection inside the loop.  Instead we return the connections to be cleaned up out and let them be destructed outside of the loop.

Reviewed By: jupyung

Differential Revision: D38913033

fbshipit-source-id: 9858303304333bafcc0516d0ed5da7666f411881