Don't generate invalid Python if field name conflicts with struct name
Summary:
Thrift's py generator generates assignments of the following form:
```
FieldName = ClassName()
```
when generating union deserializers. When `FieldName` == `ClassName`, this results in an `UnboundLocalError` at runtime when using the generated Python code.

Some users (via Configerator) seem to be creating unions which have the same field names as class names.

I think there's a valid question of whether this should be allowed/if we want to enforce standards via lint on naming of fields for consistency (seems like the cat might be out of the bag on this but I'm obviously not an expert here).

But at a minimum I'd argue we shouldn't be generating invalid Python code. This diff keeps track of local types defined and appends a less likely to conflict name only when there would otherwise be a conflict.

Reviewed By: junzh0u

Differential Revision: D39266946

fbshipit-source-id: 069907fde014571c311773bc3892c754b5e9f668