Produce field information for debuggers
Summary:
Currently the mapping between fields and their isset index is not readily available to debuggers (gdb, drgn, Object Introspection), so programmatically determining whether a field is set is not easy: it requires reading the generated code to find the index.

This diff adds minimal per-field metadata to `TStructDataStorage` that can be used to somewhat reliably inspect the struct's storage.

- The fields are private, so it not normally accessible by application code. They are only intended to be accessed by debuggers.
- The impact on compilation time should be minimal, as it only affects the `_data.cpp` file. Build sizes should also not be affected noticeably as all the string literals are already present elsewhere.
- ABI stability would be nice but not strictly necessary, as we can work around changes to the format in our tools, so technical debt should be zero.

Reviewed By: iahs

Differential Revision: D39328395

fbshipit-source-id: 617d6572408c99c8790aeb69ef2d4d0848fc2abe