require class constraints should not imply subtype
Summary:
Trait requirements usually imply subtyping: if trait `T` `require extends C`, then in the decls the elements of C are flattened inside T, and the "fake" type `T` used internally to resolve trait elements is added to the subtype relation so that `T <: C`.    This ensures that elements of `C` are seamlessly resolved when accessed from `T`.

To avoid circular dependencies when computing decls, if trait `T` `require class C`, the elements of `C` are not flattened inside `C`, but a `this = C` constraint is added when typechecking `T` and the element resolution explicitly searches for the "upper bounds on `this`".  Before this diff, despite the incomplete decls, if trait `T` `require class C`, `T <: C` was still added to the subtype relation.  When adding the constraint `this = C`, the typechecker would simplify the upper bound `this <: C & T` into `this <: T`, loosing the information that `C` elements should be explicitly searched (since they are not flattened in the decls for `T`).

This diff modifies the subtype relation, so that trait `T` `require class C`, does _not_ imply `T <: C`.  Adding the constraint `this = C` now results in the upper bound `this <: C & T` and resolution of elements works correctly.

Remark: one might wonder why tests as `call_01.good.php` used to pass.  In those tests the class `C` uses the trait `T`.  The subtype relation thus had a cycle `T <: C <: T`, and when adding the constraint `this = C` the typechecker (by chance) simplified the upper bound  `this <: C & T` into `this <: C`, enabling the search into the elements of `C`.

Reviewed By: mpu

Differential Revision: D39276533

fbshipit-source-id: 2ff0baae66d466fb8257ce17692f234819c2147f